+++ efCore-features/BookStore_Pg/Domain.Books/Entities/BookView.cs
@@ -3,7 +3,7 @@ namespace Domain.Books.Entities;
 
 public class BookView
 {
-    public int BookViewId { get; private set; }
+    public long BookViewId { get; private set; }
 
     public string Title { get; private set; }
 }
